+++ naivesearch/__init__.py
@@ -1,7 +1,5 @@
-from .indexer import InvertedIndex
 from .main import naivesearch
 
 __all__ = [
-    'InvertedIndex'
     'naivesearch'
 ]
