+++ naivesearch/indexer/__init__.py
@@ -1,8 +1,11 @@
-from .inverted_index import Chunker, InvertedIndex, Reader
+from .inverted_index import InvertedIndex, Reader
+from .preprocess_composer import Formatter, Chunker, compose_preprocessors
 
 
 __all__ = [
     'Chunker',
+    'compose_preprocessors',
+    'Formatter',
     'InvertedIndex',
     'Reader',
 ]
