+++ naivesearch/preprocessors/converter.py
@@ -1,3 +1,5 @@
+from typing import List
+
 from naivesearch.indexer import Chunker
 
 
@@ -5,6 +7,6 @@ class BigramConverter:
     def __init__(self, chunker: Chunker):
         self.chuker = chunker
 
-    def __call__(self, x: str):
+    def __call__(self, x: str) -> List[str]:
         s = self.chuker(x)
         return s + [''.join(z) for z in zip(s[0:], s[1:])]
